fix(utils): address copilot review feedback
