+++ aperturedb/CommonLibrary.py
@@ -7,6 +7,7 @@ import importlib
 import math
 import os
 import sys
+import copy
 from typing import Any, Callable, Optional, Tuple, Dict, Union
 import logging
 import json
@@ -15,6 +16,7 @@ from aperturedb.Configuration import Configuration
 from aperturedb.Connector import Connector
 from aperturedb.ConnectorRest import ConnectorRest
 from aperturedb.types import Blobs, CommandResponses, Commands
+from aperturedb.LoggingUtils import censor_tokens
 
 logger = logging.getLogger(__name__)
 
@@ -266,40 +268,6 @@ def create_connector(
     return __create_connector(config)
 
 
-def censor_tokens(data):
-    """
-    Recursively redact sensitive token fields in a dictionary or list.
-    """
-    import copy
-
-    def _censor(obj):
-        if isinstance(obj, dict):
-            for k, v in obj.items():
-                if isinstance(k, str) and k.lower() in ["refresh_token", "session_token", "token", "access_token"]:
-                    if isinstance(v, str):
-                        parts = v.split("_", 1)
-                        if len(parts) == 2:
-                            prefix = parts[0] + "_"
-                            token = parts[1]
-                        else:
-                            prefix = ""
-                            token = v
-
-                        if len(token) > 8:
-                            obj[k] = prefix + token[:4] + "..." + token[-4:]
-                        elif len(v) > 0:
-                            obj[k] = prefix + "..."
-                else:
-                    _censor(v)
-        elif isinstance(obj, list):
-            for item in obj:
-                _censor(item)
-
-    censored = copy.deepcopy(data)
-    _censor(censored)
-    return censored
-
-
 def execute_query(client: Connector, query: Commands,
                   blobs: Blobs = [],
                   success_statuses: list[int] = [0],
@@ -349,8 +317,8 @@ def execute_query(client: Connector, query: Commands,
                     raise e
     else:
         # Transaction failed entirely.
-        logger.error(f"Failed query = {query} with response = {
-                     censor_tokens(r)}")
+        logger.error(
+            f"Failed query = {query} with response = {censor_tokens(r)}")
         result = 1
 
     statuses = {}
